+++ packages/expo-router/src/testing-library/mocks.ts
@@ -1,4 +1,9 @@
-import '@testing-library/jest-native/extend-expect';
+/*
+ * Optionally enable @testing-library/jest-native/extend-expect. We use this internally for the `toBeOnTheScreen` matcher()
+ */
+try {
+  require('@testing-library/jest-native/extend-expect');
+} catch {}
 
 // include this section and the NativeAnimatedHelper section for mocking react-native-reanimated
 jest.mock('react-native-reanimated', () => {